[gtk4] -Fix: widget bgcolor compat: convert color components to int before printing with %x
git-svn-id: svn://repo.hu/librnd/trunk@33667 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/librnd/plugins/lib_gtk4_common/compat.c
@@ -379,7 +379,7 @@ void gtkc_widget_modify_bg_(GtkWidget *w, rnd_gtk_color_t *color)
 
 	rnd_snprintf(tmp, sizeof(tmp),
 		"*.wbgc {\nbackground-image: none;\nbackground-color: #%02x%02x%02x;\n}\n",
-		rnd_round(color->red*255.0), rnd_round(color->green*255.0), rnd_round(color->blue*255.0));
+		(int)rnd_round(color->red*255.0), (int)rnd_round(color->green*255.0), (int)rnd_round(color->blue*255.0));
 	gtkci_widget_css_del(w, "wbgc");
 	gtkci_widget_css_add(w, tmp, "wbgc", 0);
 }
